+++ YoutubeExplode.Tests/Data.cs
@@ -61,12 +61,14 @@ namespace YoutubeExplode.Tests
             yield return new TestCaseData("PL601B2E69B03FAB9D"); // short??
             yield return new TestCaseData("PLI5YfMzCfRtZ8eV576YoY3vIYrHjyVm_e"); // normal
             yield return new TestCaseData("PLWwAypAcFRgKFlxtLbn_u14zddtDJj3mk"); // large
-            yield return new TestCaseData("RD1hu8-y6fKg0"); // video mix
-            yield return new TestCaseData("RDMMU-ty-2B02VY"); // my mix
-            yield return new TestCaseData("RDEMNJhLy4rECJ_fG8NL-joqsg"); // music mix
+            // TODO: uncomment the following tests when a fix is found for these playlists.
+            //yield return new TestCaseData("RD1hu8-y6fKg0"); // video mix
+            //yield return new TestCaseData("RDMMU-ty-2B02VY"); // my mix
+            //yield return new TestCaseData("RDEMNJhLy4rECJ_fG8NL-joqsg"); // music mix
             yield return new TestCaseData("ULl6WWX-BgIiE"); // channel video mix
             yield return new TestCaseData("UUTMt7iMWa7jy0fNXIktwyLA"); // user uploads
             yield return new TestCaseData("PUTMt7iMWa7jy0fNXIktwyLA"); // popular user uploads
+            yield return new TestCaseData("OLAK5uy_lLeonUugocG5J0EUAEDmbskX4emejKwcM"); // music album
             yield return new TestCaseData("LLEnBXANsKmyj2r9xVyKoDiQ"); // liked
             yield return new TestCaseData("FLEnBXANsKmyj2r9xVyKoDiQ"); // favorites
         }
@@ -75,7 +77,7 @@ namespace YoutubeExplode.Tests
         {
             yield return new TestCaseData("");
             yield return new TestCaseData("PLm_3vnTS-pvmZFuF L1Pyhqf8kTTYVKjW");
-            yield return new TestCaseData("PLm_3vnTS-pvmZFuF3L1Pyhqf8kTTYVKjWz");
+            yield return new TestCaseData("PLm_3vnTS-pvmZFuF3L1Pyhqf8kTTYVKjWzWefI32jU");
             yield return new TestCaseData("PLm_3vnTS-pvmZFuF3L=Pyhqf8kTTYVKjW");
         }
 
@@ -103,6 +105,11 @@ namespace YoutubeExplode.Tests
         public static IEnumerable GetChannelIds()
         {
             yield return new TestCaseData("UCEnBXANsKmyj2r9xVyKoDiQ"); // normal
+        }
+
+        public static IEnumerable GetUsernames()
+        {
+            yield return new TestCaseData("TheTyrrr");
         }
 
         public static IEnumerable GetChannelIds_Invalid()