+++ YoutubeExplode/Models/MediaStreams/VideoStreamInfo.cs
@@ -1,4 +1,7 @@
-﻿using JetBrains.Annotations;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using JetBrains.Annotations;
 using YoutubeExplode.Internal;
 
 namespace YoutubeExplode.Models.MediaStreams
@@ -6,7 +9,7 @@ namespace YoutubeExplode.Models.MediaStreams
     /// <summary>
     /// Metadata associated with a certain <see cref="MediaStream"/> that contains only video.
     /// </summary>
-    public class VideoStreamInfo : MediaStreamInfo
+    public partial class VideoStreamInfo : MediaStreamInfo
     {
         /// <summary>
         /// Video bitrate (bits/s) of the associated stream.
@@ -44,8 +47,8 @@ namespace YoutubeExplode.Models.MediaStreams
             : base(itag, url, size)
         {
             Bitrate = bitrate.GuardNotNegative(nameof(bitrate));
-            VideoEncoding = GetVideoEncoding(itag);
-            VideoQuality = GetVideoQuality(itag);
+            VideoEncoding = ItagHelper.GetVideoEncoding(itag);
+            VideoQuality = ItagHelper.GetVideoQuality(itag);
             Resolution = resolution;
             Framerate = framerate.GuardNotNegative(nameof(framerate));
             VideoQualityLabel = VideoQuality.GetVideoQualityLabel(framerate);
@@ -57,11 +60,30 @@ namespace YoutubeExplode.Models.MediaStreams
             : base(itag, url, size)
         {
             Bitrate = bitrate.GuardNotNegative(nameof(bitrate));
-            VideoEncoding = GetVideoEncoding(itag);
+            VideoEncoding = ItagHelper.GetVideoEncoding(itag);
             Resolution = resolution;
             Framerate = framerate.GuardNotNegative(nameof(framerate));
             VideoQualityLabel = videoQualityLabel.GuardNotNull(nameof(videoQualityLabel));
-            VideoQuality = GetVideoQuality(videoQualityLabel);
+            VideoQuality = ParseVideoQualityFromLabel(videoQualityLabel);
+        }
+    }
+
+    public partial class VideoStreamInfo
+    {
+        private static readonly Dictionary<string, VideoQuality> VideoQualityLabelMap = 
+            Enum.GetValues(typeof(VideoQuality)).Cast<VideoQuality>().ToDictionary(
+                v => v.ToString().StripNonDigit(), // High1080 => 1080
+                v => v);
+
+        private static VideoQuality ParseVideoQualityFromLabel(string videoQualityLabel)
+        {
+            // Strip "p" and framerate
+            var videoQualityStr = videoQualityLabel.SubstringUntil("p");
+
+            // Try to find matching video quality
+            return VideoQualityLabelMap.TryGetValue(videoQualityStr, out var videoQuality)
+                ? videoQuality
+                : throw new FormatException($"Could not parse video quality from given string [{videoQualityLabel}].");
         }
     }
 }