+++ YoutubeExplode/Models/Playlist.cs
@@ -92,6 +92,9 @@ namespace YoutubeExplode.Models
             if (id.StartsWith("PU", StringComparison.Ordinal))
                 return PlaylistType.PopularChannelVideos;
 
+            if (id.StartsWith("OL", StringComparison.Ordinal))
+                return PlaylistType.MusicAlbum;
+
             if (id.StartsWith("LL", StringComparison.Ordinal))
                 return PlaylistType.LikedVideos;