+++ YoutubeExplode/YoutubeClient.Channel.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using AngleSharp.Dom.Html;
+using AngleSharp.Parser.Html;
 using YoutubeExplode.Exceptions;
 using YoutubeExplode.Internal;
 using YoutubeExplode.Models;
@@ -10,6 +12,38 @@ namespace YoutubeExplode
 {
     public partial class YoutubeClient
     {
+        private async Task<string> GetUserPageRawAsync(string username)
+        {
+            username = username.UrlEncode();
+            var url = $"https://www.youtube.com/user/{username}";
+            return await _httpClient.GetStringAsync(url, false).ConfigureAwait(false);
+        }
+
+        private async Task<IHtmlDocument> GetUserPageAsync(string username)
+        {
+            var raw = await GetUserPageRawAsync(username).ConfigureAwait(false);
+            return await new HtmlParser().ParseAsync(raw).ConfigureAwait(false);
+        }
+
+        /// <inheritdoc />
+        public async Task<string> GetChannelIdAsync(string username)
+        {
+            username.GuardNotNull(nameof(username));
+
+            // Get user page
+            var userPage = await GetUserPageAsync(username).ConfigureAwait(false);
+
+            // Extract channel ID
+            var channelId = userPage.QuerySelector("link[rel=\"canonical\"]").GetAttribute("href")
+                .SubstringAfter("channel/");
+
+            // Validate channel ID
+            if (!ValidateChannelId(channelId))
+                throw new ParseException("Could not parse channel ID.");
+
+            return channelId;
+        }
+
         /// <inheritdoc />
         public async Task<Channel> GetChannelAsync(string channelId)
         {